Retrieve principal from front end
