+++ kies/pom.xml
@@ -163,11 +163,6 @@
 			<artifactId>spring-security-config</artifactId>
 			<version>5.6.2</version>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-oauth2-client</artifactId>
-			<version>5.6.2</version>
-		</dependency>
 	</dependencies>
 
 	<build>
