+++ kies/src/main/java/com/charlotte/kies/repository/UserRepository.java
@@ -12,4 +12,5 @@ import java.util.List;
 public interface UserRepository extends JpaRepository<User, Long> {
     List<User> findByEmail(String emailAddress);
 
+    User getUserByUsername(String username);
 }
