+++ kies/src/main/java/com/charlotte/kies/service/UserService.java
@@ -1,0 +1,30 @@
+package com.charlotte.kies.service;
+
+import com.charlotte.kies.Provider;
+import com.charlotte.kies.model.User;
+import com.charlotte.kies.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+
+    @Autowired
+    private UserRepository repo;
+
+
+
+    public void processOAuthPostLogin(String username) {
+        User existUser = repo.getUserByUsername(username);
+
+        if (existUser == null) {
+            User newUser = new User();
+            newUser.setUsername(username);
+            newUser.setProvider(Provider.GOOGLE);
+            newUser.setEnabled(true);
+
+            repo.save(newUser);
+        }
+
+    }
+}
