+++ kies/src/main/resources/application.properties
@@ -62,3 +62,8 @@ spring.security.oauth2.resourceserver.jwt.issuer-uri: https://accounts.google.co
 server.session.persistent=true
 
 server.servlet.session.persistent=true
+
+spring.security.oauth2.client.registration.google.client-id: {CLIENT_ID}
+spring.security.oauth2.client.registration.google.client-secret: {CLIENT_SECRET}
+spring.security.oauth2.client.registration.google.redirectUri: "{baseUrl}/oauth2/callback/{registrationId}"
+spring.security.oauth2.client.registration.google.scope: email,profile
