+++ kies/src/main/resources/db/changelog/add-enabled-column.sql
@@ -1,0 +1,10 @@
+--liquibase formatted sql
+
+--changeset charlotte.kies:7
+
+
+ALTER TABLE users
+   ADD COLUMN enabled boolean;
+
+
+--rollback ALTER TABLE users DROP COLUMN enabled;
